% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.R
\name{np}
\alias{np}
\title{Tamaño de la muestra para la estimación de una proporción}
\usage{
np(z, p, e)
}
\arguments{
\item{z}{valor del percentil normal que indica el nivel de confianza}

\item{p}{valor de la proporcion estimada}

\item{e}{valor del error de muestreo (diferencia entre la media muestral y el parámetro)}
}
\value{
tamaño de muestra para la estimación de una proporción
}
\description{
Función que calcula el tamaño de la muestra para la estimación de un media
}
\examples{
 m.mu(1.96,0.5,0.02) 
}
