% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartest.R
\name{vartes}
\alias{vartes}
\title{Prueba chi-cuadrado para la varianza de una poblacion}
\usage{
vartes(x, sigma0, alternative, conf.level)
}
\arguments{
\item{x}{vector con los datos}

\item{sigma0}{valor a contrastar}

\item{alternative}{tipo de prueba : c("two.sided", "less", "greater")}

\item{conf.level}{nivel de confianza}
}
\value{
valorp
}
\description{
Función que calcula el valor p para una prueba de varianza
}
\examples{
 x=rnorm(100,120,4)   
           vartest(x, sigma0=20,alternative ="two.sided", conf.level=0.95)
}
